Rename useTemplates to useTemplatesApi
